CICD_stocks_products/requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
psycopg2-binary
python-decouple
django-filter
gunicorn~=20.1.0
pytest-django
flake8
pytest
django-heroku